Refactor abort-on-fail test
